backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
openai
pydantic
python-dotenv
